Volumetric Clouds!
This is happening! My first ever volumetric clouds! Credit to RRe36 for really really helping me with the code! Hope you all like the clouds! The description of the CloudsType is updated so read it to know how to make the Volumetric Clouds work in the best way! (You don't need to change the code) and more optimaztion options for the clouds soon!